+++ CHANGELOG.md
@@ -11,10 +11,10 @@
 ### New features
 
 * New option in preprocess to check that sizes of source and target are equal (for seqtagging)
-* visual beam dumper option
 * Add `-pdbrnn_merge` option to define how to reduce the time dimension
 * Add option to segment mixed cased words
 * Add options to segment words of given alphabets or when switching alphabets
+* Add option to save the beam search as JSON
 
 ### Fixes and improvements